+++ core/src/androidTest/java/org/hisp/dhis/android/core/common/CategoryMother.java
@@ -7,7 +7,7 @@ import org.hisp.dhis.android.core.category.Category;
 
 import java.util.Date;
 
-public class CategoryMockFactory {
+public class CategoryMother {
 
     public static Category generateCategory(@NonNull String uid) {
         Date today = new Date();
